geocode-place: Fix -Wswitch-enum warnings
This switch is not meant to handle all of the values for the switch
(omissions are not a bug), so cast the value to an int to disable type
checking first.

https://bugzilla.gnome.org/show_bug.cgi?id=756311

+++ geocode-glib/geocode-place.c
@@ -1127,7 +1127,7 @@ get_icon_name (GeocodePlace *place)
 {
         const char *icon_name;
 
-        switch (place->priv->place_type) {
+        switch ((int) place->priv->place_type) {
 
         case GEOCODE_PLACE_TYPE_BUILDING:
                 icon_name = "poi-building";
